+++ clean.sh
@@ -1,2 +1,0 @@
-rm -rf out 2> /dev/null
-rm -rf types 2> /dev/null